+++ src/babashka/impl/classes.clj
@@ -13,7 +13,8 @@
 ;;                     (identical? os :mac)))
 
 (def classes
-  '{:all [java.io.BufferedReader
+  '{:all [clojure.lang.ExceptionInfo
+          java.io.BufferedReader
           java.io.BufferedWriter
           java.io.ByteArrayInputStream
           java.io.ByteArrayOutputStream
@@ -21,6 +22,8 @@
           java.io.InputStream
           java.io.IOException
           java.io.OutputStream
+          java.io.FileReader
+          java.io.PushbackInputStream
           java.io.Reader
           java.io.StringReader
           java.io.StringWriter
@@ -28,6 +31,7 @@
           java.lang.ArithmeticException
           java.lang.AssertionError
           java.lang.Boolean
+          java.lang.Byte
           java.lang.Class
           java.lang.Double
           java.lang.Exception
@@ -55,6 +59,8 @@
           java.net.URLDecoder
           java.nio.file.CopyOption
           java.nio.file.FileAlreadyExistsException
+          java.nio.file.FileSystem
+          java.nio.file.FileSystems
           java.nio.file.Files
           java.nio.file.LinkOption
           java.nio.file.NoSuchFileException
@@ -109,8 +115,7 @@
     :methods [borkdude.graal.LockFix ;; support for locking
               ]
     :fields [clojure.lang.PersistentQueue]
-    :instance-checks [clojure.lang.ExceptionInfo
-                      clojure.lang.IObj
+    :instance-checks [clojure.lang.IObj
                       clojure.lang.IEditableCollection]
     :custom {clojure.lang.LineNumberingPushbackReader {:allPublicConstructors true
                                                        :allPublicMethods true}
@@ -200,7 +205,11 @@
                    (instance? java.security.MessageDigest v)
                    java.security.MessageDigest
                    (instance? java.io.InputStream v)
-                   java.io.InputStream)))))
+                   java.io.InputStream
+                   (instance? java.io.OutputStream v)
+                   java.io.OutputStream
+                   (instance? java.nio.file.FileSystem v)
+                   java.nio.file.FileSystem)))))
 
 (def class-map (gen-class-map))