+++ src/babashka/impl/repl.clj
@@ -11,11 +11,15 @@
    [sci.impl.parser :as parser]
    [sci.impl.vars :as vars]))
 
+(set! *warn-on-reflection* true)
+
 (defn repl-caught
   "Default :caught hook for repl"
-  [e]
+  [^Throwable e]
   (sci/with-bindings {sci/out @sci/err}
-    (sio/println (.getMessage ^Exception e))
+    (sio/println (str (.. e getClass getName)
+                      (when-let [m (.getMessage e)]
+                        (str ": " m)) ))
     (sio/flush)))
 
 (defn repl